Restructured files, and begun creation of new, prettier, homepage.
