+++ audioFunctions.py
