+++ templates/home.html
@@ -1,0 +1,45 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>Mutify</title>
+    <link rel="stylesheet" type="text/css" href="style.css">
+    <script src="tabscript.js"></script>
+  </head>
+  <body>
+    <h1>Mutify</h1>
+    <section class="container">
+      <div class="projectinfo">
+	Mofu mofu!
+      </div>
+      <div class="upload">
+	<! -- Tab Links: -->
+	<div class="tab">
+	  <button class="tabbutton" onclick="openTab(event, 'Split')">Split</button>
+	  <button class="tabbutton" onclick="openTab(event, 'Crop')">Crop</button>
+	  <button class="tabbutton" onclick="openTab(event, 'SpeedUp')">Speed Up</button>
+	  </div>
+	<! -- Tab Content: -->
+	<div id="Split" class="tabcontent">
+	  <h4>Upload Album</h4>
+	  <form action="/" method="post" enctype="multipart/form-data">
+	    Album (Please have timestamps and tracknames separated by newlines):<br>
+	    <input type="text" value="Album Name" name="album" required="required" />
+	    <input type="text" value="Artist Name" name="artist" required="required" />
+	    <input type="file" name="albumfile" required="required" />
+	    <br>Time Stamps:<br>
+	    <p>
+	      <textarea name="timestamps" cols="60" rows="10" required="required"></textarea><br>
+	      <br>Track Names:<br>
+	      <textarea name="tracknames" cols="60" rows="10" required="required"></textarea><br>
+	      <input type="submit" value="Submit"> <input type="reset"></p>
+	  </form>
+	</div>
+	<div id="Crop" class="tabcontent">
+	  Crop Content.
+	</div>
+	<div id="SpeedUp" class="tabcontent">
+	  Speed Up Content.
+	  </div>
+      </div>
+  </body>
+</html>
