--- templates/out.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <style>
-      table {
-	  font-family: calibri, sans-serif;
-	  border-collapse: collapse;
-	  width: 100%;
-      }
-      td, th {
-	  border: 1px solid #ddddd;
-	  text-align: left;
-	  padding: 8px;
-      }
-    </style>
-  </head>
-  <body>
-    <h2>Tracks</h2>
-    <div>
-      <h3>
-	Mutify:
-	<a href = "outzip">Download as zip</a>
-      </h3>
-    </div>
-    <table>
-      <tr>
-	<th>Track Name</th>
-      </tr>
-      {% for key in tracks %}
-      <tr>
-	<td><a href="/out/{{key}}">{{key}}.mp3</a></td>
-      </tr>
-      {% endfor %}
-    </table>
-  </body>
-</html>
